Admin GUI functionality Complete
Dev Notes:
Please delete and remake your databases with the new dbManager

Added the ability to add a new offering/ course to the catalog.
Added error checking to ensure the course is added to the database correctly
Database code updated for a minor redundancy
TODO:
Error checking on all components of project.
JavaDoc on my code and others
Potentially J-Unit testing on classes